+++ app/item.py
@@ -1,0 +1,7 @@
+class item:
+  alt_name   = ""
+  typeID     = "0"
+  alt_typeID = "0"
+  def __init__(self, name, amount):
+    self.name     = name
+    self.amount   = amount
