+++ app/order.py
@@ -1,0 +1,8 @@
+class order:
+  def __init__(self, name, buy, sell, volume, buy_corp, sell_corp):
+    self.name      = name
+    self.buy       = buy
+    self.sell      = sell
+    self.volume    = volume
+    self.buy_corp  = buy_corp
+    self.sell_corp = sell_corp
